requirements.txt
absl-py==2.2.2
addict==2.4.0
audioread==3.0.1
basicsr==1.4.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
decorator==5.2.1
face-alignment==1.3.5
facexlib==0.3.0
ffmpeg-python==0.2.0
filelock==3.18.0
filterpy==1.4.5
fonttools==4.57.0
fsspec==2025.3.2
future==1.0.0
gfpgan==1.3.8
grpcio==1.71.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.4.7
importlib_metadata==8.6.1
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.1.0
kiwisolver==1.4.7
kornia==0.6.8
lazy_loader==0.4
librosa==0.9.2
llvmlite==0.43.0
lmdb==1.6.2
Markdown==3.8
MarkupSafe==3.0.2
matplotlib==3.9.4
mpmath==1.3.0
networkx==3.2.1
numba==0.60.0
numpy==1.23.4
opencv-python==4.11.0.86
packaging==25.0
pillow==11.2.1
platformdirs==4.3.7
pooch==1.8.2
protobuf==6.30.2
pycparser==2.22
pydub==0.25.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
resampy==0.3.1
safetensors==0.5.3
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.10.1
six==1.17.0
soundfile==0.13.1
sympy==1.13.3
tb-nightly==2.20.0a20250424
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tifffile==2024.8.30
tomli==2.2.1
torch==1.12.1+cu113
torchaudio==0.12.1+cu113
torchvision==0.13.1+cu113
tqdm==4.67.1
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
yacs==0.1.8
yapf==0.43.0
zipp==3.21.0
